refactor: clippy to zero, and the ceiling is gone
Issue #5, the owner's instruction: "This needs to be 0." It is 0, on both
toolchains, and `tools/clippy.sh` is now `-D warnings` with no ceiling and no
`--update`. The two `tools/clippy-ceiling-*.txt` files are deleted. A single
finding now fails the run, the same way a rustc warning does.

Removing the number was the point, not lowering it. A count in a file has to be
maintained by hand every time someone improves something, and it reads as an
allowance rather than a standard. The ceiling was a holding position while there
were 161 findings and none of them a bug; it held until the count reached zero,
and then it only created work.

THE PER-PLATFORM CEILING WAS REALLY PER-CLIPPY-VERSION, and that mattered more
than the number did:

    this laptop, Homebrew   clippy 0.1.95    45 findings
    build host and CI       clippy 0.1.98   123 findings

Identical source, 78 apart. New lints arrive with new versions, and cfg-gated
code is only linted where it compiles, so `watcher.rs`'s kqueue block is
invisible to clippy on Linux and the Linux-only lints are invisible here. The
practical consequence is that `--fix` had to run on the BUILD HOST, with the
changed sources pulled back and the diff read here under git. Fixing from a 1.95
laptop would have been fixing the wrong list, and the ceiling files had been
hiding that for as long as they existed.

WHAT IT ACTUALLY TOOK. The breakdown recorded in HANDOVER.md was taken at 135
and had rotted like every other number in that file: it promised 18 collapsible
`if` and 30 auto-fixable, and the real shape was different.

  --fix cleared 123 to 49, all behaviour-preserving. The previous commit.

  The other 49 wanted decisions, and the owner's call was to reshape everything
  and allow nothing:
    - type aliases for the `Arc<dyn Fn(..)>` builder shapes: InitGlobal,
      InitThread and Destroy in app.rs, CachedErrorPage and FinalizedResponse in
      m6-http, TokenRow in m6-auth, and a named UaAcc struct in telemetry.rs
      replacing a `(u64, HashMap, HashMap)` that was read positionally as `e.1`
      and `e.2` with nothing saying which held addresses and which held paths
    - `is_empty` beside `len` on the cache and the route table
    - four oversized enum variants boxed. `ConnKind::Http2` is the one worth
      arguing: 336 bytes against 80, so inline it sized every stored `Conn`,
      H1 and mid-handshake included. The cost is one allocation per HTTP/2
      connection at ALPN selection, not per request, and connections here are
      long-lived. On a 950MB single-core origin holding a map of these, 256
      bytes saved on every other entry is the better side of that trade.
    - `&mut Vec<u8>` to `&mut [u8]` where the buffer is only ever resliced
    - two argument lists grouped into EventLoopIo and H2Response, taking an
      8-argument and a 9-argument function to 4 and 6. Both are honest
      groupings rather than a bag to satisfy a lint: the five EventLoopIo
      fields are exactly what the loop polls, and the four H2Response fields
      are the response. Destructured at the top of each function so the bodies
      are untouched.

THREE WERE REAL DOCUMENTATION BUGS, not style. `health.rs` carried the doc
comment for a timing-safe credential comparison directly above
`#[cfg(test)] mod token_file_tests`, so it documented the test module. The
function it describes had moved to `m6-core/src/monitoring.rs` during the
consolidation and the comment stayed behind, byte-identical to the one now
correctly attached there. Deleted. Two module-level prose blocks, in `http11.rs`
and `hints.rs`, were `///` rather than `//!`, so they documented the following
`use` statement instead of the module.

ONE WAS ARGUABLY A BUG: `drop(req)` in `m6-core/src/h1.rs`, commented "release
borrow of `headers`". It compiles without it, so NLL had been ending that borrow
on its own and the line was pre-NLL residue promising something it did not do.

The two `unnecessary_cast` findings in `watcher.rs` were checked against libc
rather than assumed, because that file is cfg-gated to macos, freebsd and
openbsd and only macOS is ever built: `NOTE_WRITE` is `u32` and `fflags` is
`u32` or `c_uint` on all three, so the casts were redundant everywhere the cfg
applies and removing them cannot break the platforms nothing tests.

Verified: clippy silent on BOTH, measured after `cargo fmt` rather than before,
because formatting touched files `--fix` had rewritten. 0.1.95 on this laptop
and 0.1.98 on the build host. 1022 tests passed, 0 failed. 0 rustc warnings on
release and test builds on Linux. `cargo fmt --check` clean. Full build-host
gate runs before this merges.

CI's step and the site repo's run-tests.sh header both described a count that
may fall and never rise, and now describe what they actually enforce.

Closes #5

+++ m6-http/src/watcher.rs
@@ -257,12 +257,13 @@ impl FsWatcher {
 
                 if (event.mask.contains(EventMask::DELETE)
                     || event.mask.contains(EventMask::MOVED_FROM))
-                    && name.ends_with(".sock") {
-                        result.push(FsEvent {
-                            path: self.inner.socket_dir.join(&name),
-                            kind: FsEventKind::SocketDeleted,
-                        });
-                    }
+                    && name.ends_with(".sock")
+                {
+                    result.push(FsEvent {
+                        path: self.inner.socket_dir.join(&name),
+                        kind: FsEventKind::SocketDeleted,
+                    });
+                }
 
                 if event.mask.contains(EventMask::CLOSE_WRITE) {
                     if name == "site.toml" || name.is_empty() {
@@ -364,7 +365,7 @@ fn kqueue_watch_site_dir(site_dir: PathBuf, pipe_write: RawFd) {
         ident: dir_fd as libc::uintptr_t,
         filter: libc::EVFILT_VNODE,
         flags: libc::EV_ADD | libc::EV_ENABLE | libc::EV_CLEAR,
-        fflags: (libc::NOTE_WRITE | libc::NOTE_EXTEND | libc::NOTE_ATTRIB | libc::NOTE_LINK) as u32,
+        fflags: libc::NOTE_WRITE | libc::NOTE_EXTEND | libc::NOTE_ATTRIB | libc::NOTE_LINK,
         data: 0,
         udata: std::ptr::null_mut(),
     };
@@ -383,8 +384,7 @@ fn kqueue_watch_site_dir(site_dir: PathBuf, pipe_write: RawFd) {
             ident: file_fd as libc::uintptr_t,
             filter: libc::EVFILT_VNODE,
             flags: libc::EV_ADD | libc::EV_ENABLE | libc::EV_CLEAR,
-            fflags: (libc::NOTE_WRITE | libc::NOTE_ATTRIB | libc::NOTE_RENAME | libc::NOTE_DELETE)
-                as u32,
+            fflags: libc::NOTE_WRITE | libc::NOTE_ATTRIB | libc::NOTE_RENAME | libc::NOTE_DELETE,
             data: 0,
             udata: std::ptr::null_mut(),
         };
